+++ splunk_otel/distro.py
@@ -19,11 +19,15 @@ from typing import Any, Dict
 from opentelemetry.instrumentation.distro import BaseDistro  # type: ignore
 from pkg_resources import EntryPoint
 
+from splunk_otel.metrics import _configure_metrics
 from splunk_otel.options import _Options
-from splunk_otel.tracing import _configure_tracing, _is_truthy
+from splunk_otel.tracing import _configure_tracing
+from splunk_otel.util import _is_truthy
+
+otel_log_level = os.environ.get("OTEL_LOG_LEVEL", logging.INFO)
 
 logger = logging.getLogger(__file__)
-logger.setLevel(logging.INFO)
+logger.setLevel(otel_log_level)
 
 
 class _SplunkDistro(BaseDistro):
@@ -35,9 +39,18 @@ class _SplunkDistro(BaseDistro):
                 "tracing has been disabled with OTEL_TRACE_ENABLED=%s", tracing_enabled
             )
 
+        metrics_enabled = os.environ.get("OTEL_METRICS_ENABLED", True)
+        self._metrics_enabled = _is_truthy(metrics_enabled)
+        if not self._metrics_enabled:
+            logger.info(
+                "metering has been disabled with OTEL_METRICS_ENABLED=%s", metrics_enabled
+            )
+
     def _configure(self, **kwargs: Dict[str, Any]) -> None:
         if self._tracing_enabled:
             _configure_tracing(_Options())
+        if self._metrics_enabled:
+            _configure_metrics()
 
     def load_instrumentor(self, entry_point: EntryPoint, **kwargs):
         if self._tracing_enabled: